pyproject.toml
[tool.poetry]
name = "invera-todo-challenge"
version = "0.1.0"
description = "Simple TODO API"
authors = ["Franco Pelusso <franco.1.gm@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
Django = "^3.2.7"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
tox = "^3.24.4"
flake8 = "^3.9.2"
pylint = "^2.11.1"
pytest-django = "^4.4.0"
pylint-django = "^2.4.4"
pytest-cov = "^2.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.master]
ignore = "migrations"
load-plugins = "pylint_django"
django-settings-module = "todo_challenge.settings"

[tool.pylint.format]
max-line-length = 120
